Managing GAP packages

The following functions allow one to load/install/update/remove GAP packages.

GAP.Packages.load — Function

load(spec::String, version::String = ""; install = false)

Try to load the newest installed version of the GAP package with name spec. Return true if this is successful, and false otherwise.

The function calls GAP's LoadPackage function; the package banner is not printed.

If install is set to true and the required GAP package is not yet installed then install is called first, in order to install the newest released version of the package.

GAP.Packages.install — Function

install(spec::String; interactive::Bool = true, quiet::Bool = false)

Download and install the newest released version of the GAP package given by spec in the pkg subdirectory of GAP's build directory (variable GAP.GAPROOT). Return true if the installation is successful or if the package was already installed, and false otherwise.

spec can be either the name of a package or the URL of an archive or repository containing a package, or the URL of a PackageInfo.g file.

The function uses the function InstallPackage from GAP's package PackageManager. The info messages shown by this function can be suppressed by entering true as the value of quiet.

GAP.Packages.update — Function

update(spec::String; quiet::Bool = false)

Update the GAP package given by spec that is installed in the pkg subdirectory of GAP's build directory (variable GAP.GAPROOT), to the latest version. Return true if a newer version was installed successfully, or if no newer version is available, and false otherwise.

spec can be either the name of a package or the URL of an archive or repository containing a package, or the URL of a PackageInfo.g file.

The function uses the function UpdatePackage from GAP's package PackageManager. The info messages shown by this function can be suppressed by entering true as the value of quiet.

GAP.Packages.remove — Function

remove(spec::String; quiet::Bool = false)

Remove the GAP package with name spec that is installed in the pkg subdirectory of GAP's build directory (variable GAP.GAPROOT). Return true if the removal was successful, and false otherwise.

The function uses the function RemovePackage from GAP's package PackageManager. The info messages shown by this function can be suppressed by entering true as the value of quiet.