GAP.jl

GAP.jl

GAP.jl

Introduction

Types

Macros

Conversions

Convenience adapters

Managing GAP packages

Other

Index

Introduction

TODO: describe the goals and non-goals of this package

TODO: also link to JuliaInterface docs (also JuliaExperimental?)

Types

GAP.FFE — Type.

FFE

Wraps a pointer to a GAP FFE immediate object. This type is defined in the JuliaInterface C code.

GAP.GapObj — Type.

GapObj

TODO

Macros

GAP.@gap — Macro.

@gap <expr>
@gap(<expr>)

Execute <expr> directly in GAP, as if GAP.EvalString("<expr>") was called. This can be used for creating GAP literals directly from Julia.

Examples

julia> @gap [1,2,3]
GAP: [ 1, 2, 3 ]

julia> @gap SymmetricGroup(3)
GAP: Sym( [ 1 .. 3 ] )

julia> @gap(SymmetricGroup)(3)
GAP: Sym( [ 1 .. 3 ] )

Note that the last two examples have a slight syntactical, and therefore also a semantical difference. The first one executes the string SymmetricGroup(3) directly inside GAP. The second example returns the function SymmetricGroup via @gap(SymmetricGroup), then calls that function with the argument 3.

Due to Julia's way of handing over arguments into the code of macros, not all expressions representing valid GAP code can be processed. For example, the GAP syntax of permutations consisting of more than one cycle cause problems, as well as the GAP syntax of non-dense lists.

julia> @gap (1,2,3)
GAP: (1,2,3)

julia> @gap (1,2)(3,4)
ERROR: LoadError: Error thrown by GAP: Error, no method found! For debugging hints type ?Recovery from NoMethodFound
[...]

julia> @gap [ 1,, 2 ]
ERROR: syntax: unexpected ","
[...]

Note also that a string argument gets evaluated with GAP.EvalString.

julia> @gap "\"abc\""
GAP: "abc"

julia> @gap "[1,,2]"
GAP: [ 1,, 2 ]

julia> @gap "(1,2)(3,4)"
GAP: (1,2)(3,4)

GAP.@g_str — Macro.

@g_str

Create a GAP string by typing g"content".

Examples

julia> g"foo"
GAP: "foo"

Conversions

GAP.gap_to_julia — Function.

gap_to_julia(type, x, recursion_dict=nothing; recursive=true)

Tries to convert the object x to a Julia object of type type. If x is a GAP.GapObj then the conversion rules are defined in the manual of the GAP package JuliaInterface. If x is another GAP.Obj (for example a Int64) then the result is defined in Julia by type.

The parameter recursion_dict is meant to preseve the identity of converted subobjects and should never be given by the user.

For GAP lists and records, it makes sense to convert also the subobjects recursively, or to keep the subobjects as they are; the behaviour is controlled by recursive, which can be true or false.

GAP.julia_to_gap — Function.

julia_to_gap(input, recursive::Val{Recursive} = Val(false), recursion_dict = IdDict())

Converts a julia object input to an appropriate GAP object. If recursive is set to Val(true), recursive conversions on arrays, tuples, and dictionaries is performed.

The input recursive_data should never be set by the user, it is meant to keep egality of input data, by converting egal data to identical objects in GAP.

Convenience adapters

TODO: describe the various convenience / adapter methods we install, e.g. for basic arithmetic, accessing GAP list and record entries, calling GAP function, etc.

TODO: Describe access to arbitrary GAP variables and functions via GAP.Globals.IDENTIFIER_NAME

TODO: describe Help system integration

Managing GAP packages

GAP.Packages.load — Function.

load(desc)

TODO: write text, link to GAP's LoadPackage

TODO: support version qualifier?

GAP.Packages.install — Function.

install(desc)

TODO: copy text from https://gap-packages.github.io/PackageManager/doc/chap2.html resp. link to that

GAP.Packages.update — Function.

update(desc)

TODO: copy text from https://gap-packages.github.io/PackageManager/doc/chap2.html resp. link to that

GAP.Packages.remove — Function.

remove(desc)

TODO: copy text from https://gap-packages.github.io/PackageManager/doc/chap2.html resp. link to that

Other

GAP.prompt — Function.

GAP.prompt()

Start a GAP prompt where you can enter GAP commands as in a regular GAP session. This prompt can be left as any GAP prompt by either entering quit; or pressing ctrl-D, which returns to the Julia prompt.

This GAP prompt allows to quickly switch between writing Julia and GAP code in a session where all data is shared.

Index

GAP.FFE

GAP.GapObj

GAP.Packages.install

GAP.Packages.load

GAP.Packages.remove

GAP.Packages.update

GAP.gap_to_julia

GAP.julia_to_gap

GAP.prompt

GAP.@g_str

GAP.@gap