GAP.jl

GAP.jl

GAP.jl

Introduction

Types

Macros

Conversions

Convenience adapters

Index

Introduction

TODO: describe the goals and non-goals of this package

TODO: also link to JuliaInterface docs (also JuliaExperimental?)

Types

GAP.FFE — Type.

FFE

Wraps a pointer to a GAP FFE immediate object. This type is defined in the JuliaInterface C code.

GAP.GapObj — Type.

GapObj

TODO

Macros

GAP.@gap — Macro.

@gap <expr>
@gap(<expr>)

Execute <expr> directly in GAP, as if GAP.EvalString("<expr>") was called. This can be used for creating GAP literals directly from Julia.

Examples

julia> @gap [1,2,3]
GAP: [ 1, 2, 3 ]

julia> @gap SymmetricGroup(3)
GAP: Sym( [ 1 .. 3 ] )

julia> @gap(SymmetricGroup)(3)
GAP: Sym( [ 1 .. 3 ] )

Note that the last two examples have a slight syntactical, and therefore also a semantical difference. The first one executes the string SymmetricGroup(3) directly inside GAP. The second example returns the function SymmetricGroup via @gap(SymmetricGroup), then calls that function with the argument 3.

Due to Julia's way of handing over arguments into the code of macros, not all expressions representing valid GAP code can be processed. For example, the GAP syntax of permutations consisting of more than one cycle cause problems, as well as the GAP syntax of non-dense lists.

julia> @gap (1,2,3)
GAP: (1,2,3)

julia> @gap (1,2)(3,4)
ERROR: LoadError: Error thrown by GAP: Error, no method found! For debugging hints type ?Recovery from NoMethodFound
[...]

julia> @gap [ 1,, 2 ]
ERROR: syntax: unexpected ","
[...]

Note also that a string argument gets evaluated with GAP.EvalString.

julia> @gap "\"abc\""
GAP: "abc"

julia> @gap "[1,,2]"
GAP: [ 1,, 2 ]

julia> @gap "(1,2)(3,4)"
GAP: (1,2)(3,4)

GAP.@g_str — Macro.

@g_str

Create a GAP string by typing g"content".

Examples

julia> g"foo"
GAP: "foo"

Conversions

Missing docstring.

Missing docstring for gap_to_julia(). Check Documenter's build log for details.

Missing docstring.

Missing docstring for julia_to_gap(). Check Documenter's build log for details.

Convenience adapters

TODO: describe the various convenience / adapter methods we install, e.g. for basic arithmetic, accessing GAP list and record entries, calling GAP function, etc.

TODO: Describe access to arbitrary GAP variables and functions via GAP.Globals.IDENTIFIER_NAME

TODO: describe Help system integration

Index

GAP.FFE

GAP.GapObj

GAP.@g_str

GAP.@gap